Initial commit for Ada++
This is the initial commit for Ada++ focusing mostly on keyword additions,
modifications to the syntax and some subtype additions to Standard. Also,
initial support of implicit with has been added as well. Note, when compiling
with GNAT mode -gnatg the new keywords are disabled.

There are currently no modificatino in how errors are reported to reflect
syntactic additions.